Change from long to long long.

+++ Levenshtein/_levenshtein.h
@@ -103,8 +103,8 @@ lev_u_edit_distance(size_t len1,
                     int xcost);
 
 _LEV_STATIC_PY size_t
-lev_l_edit_distance(size_t len1, const long *string1,
-                    size_t len2, const long *string2,
+lev_l_edit_distance(size_t len1, const long long *string1,
+                    size_t len2, const long long *string2,
                     int xcost);
 
 _LEV_STATIC_PY
